third_party/WebKit/LayoutTests/fast/selectors/style-sharing-shadow-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
PASS true is true
PASS true is true
PASS false is false
PASS false is false
PASS false is false
PASS true is true
PASS false is false
PASS false is false
item 1
item 2
A

B
